Use attr_reader
When setting instance variable that will be used mostly for reading, we
usually put them as `attr_reader` attribute. This allows to access the
variable without `@`, as if it were a method.

The reason for that is that it's hard to track the value of an instance
variable. If it's a variable, we expect its value will change. What is
its value when you write a new method that use it? Is it set at all?
(this is a real question, because you can access any
@unexisting_variable, it just returns nil).

This confusion is especially not necessary since you usually just set
the variable in the class constructor, then keep it unchanged: you just
want to access a value that come from outside.

Using `attr_reader`, a developer reading your code know for sure the
value won't change, and can easily find where it's set.