Use subclasses
Having a lot of methods having the same prefix is usually a code smell
meaning you should use a subclass (the same applies when you have a lot
of methods that take the same arguments).

This indicates they represent an entity, which you can extract as a
class to make it more clear to a developer reading your code what
elements compose your feature.