Split up big methods
When methods are getting big, this usually means it could be composed
using several smaller methods.

This will make it way easier to read: "ok, running is about purging
processes, or accept connection if there is nothing to purge". The
bigger your method will be, the more painful it will be to read. Be a
nice guy and add intuitive names on your logic chunks :)

Object Oriented programming is all about feeling what is too big or out
of place. If there are "too many" lines in a method, it should be
several methods. If there are too many methods in a class, it should be
several (sub)classes. If there are too many classes in a project, it
should be several libs.

Note that in this diff, the subclass has been put under `private`. This
does not affects classes, you can put them just where you want (I like
to put them at the bottom of the class, but it's just a personal taste).

When you split up methods, it's better to put the newly created methods
under the private scope, so that someone reading your code know it's
just implementation details and is not supposed to be used from outside.