Map fs to class hierarchy
This is not especially mandatory for such a simple project, but if you
intend to grow it, you need to organize your files.

It's idiomatic in ruby to have a simple file that requires all needed
files and define a few constants, like the version, which allow for
easier debugging when someone comes to you with a problem.